+++ teste/SME.SGP.TesteIntegracao/Setup/appsettings.json
@@ -1,0 +1,3 @@
+{
+  "UrlFrontEnd": "http://localhost:3000"
+}
